requirements.txt
flask>=3.0
flask-socketio>=5.3.6
eventlet>=0.36.1
tensorflow==2.15.*
numpy>=1.24,<2.0
pillow>=10.2.0
scikit-learn>=1.4
matplotlib>=3.7
pandas>=2.0
qrcode>=7.4
firebase-admin>=6.5.0
python-dotenv>=1.0
